Added regcontext number matching:
When a number is dialed, it is also match against the regcontext dialplan entries, which are created for every SCCP phone when it registers at an asterisk server.
This context never contains any patterns to be matched, which should proclude pattern matching and possible faults, thereof.

For now this feature is only active when compiled --enable-advanced-function.

check: "show dialplan sccpregistration" to see the registered numbers



git-svn-id: https://chan-sccp-b.svn.sourceforge.net/svnroot/chan-sccp-b/trunk@2383 fa19cda8-b258-4bce-aa1a-843e42da0b19