Fix: revised/refactored the scheduled digittimeout, hangup system. Fixes:refcount issue created by fix in 6384, which revealed the schedule issue.
Note: Because of the returned value of ast_sched_add_ref (-1 for now schedule and > 0 for a new schedule_id), most of the checks needed to be fixed (as well). 
Scheduler id's start at 0, which negated most of the checks when testing for it, delivering different results on the first call (on newly started astereisk)
then compared to later calls.



git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@6388 43c1dc8c-776e-4df2-8ef1-6e829299be21