Enhancement: Interpret channel variable:SCCP_VIDEO_MODE (off/user/auto)
When CHANNEL(video=???) is set it will also add variable '_SCCP_VIDEO_MODE'
with the same value to the call, so that this state can be picked up by the
other side.

Also added Dial option video=??? as in Dial(SCCP/1234/video=off,20,W), to only
change the video mode on one side of the call.

This will allow you to activate or suppress video both on incoming and outcoming
calls by using one of these additional dialplan options.

git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@6398 43c1dc8c-776e-4df2-8ef1-6e829299be21