Fix: setting CHANNELSTATE_CONNECTED should be reentrant
Issue: going from CONNECTED to CONNECTED (to update the callerid information) was not checking the rtp instance correctly,
causing another call to openReceiveChannel, and the previous openReceiveChannel was not stopped or cleaned up causing
the phone to run out of RTP ports. Reported by Miguel.

Solution: fix the check before calling openReceiveChannel.

related to: #297