Fix: crossover session causing re-registration trouble and refcount issues
We are cleaning up the device from the new thread, if/when the server then starts to
cleanup the device from the old thread as well, everything goes really wrong.
Instead of cleaning up the device from the new thread (and potentially from the old
thread (keepalive timeout > 60) at the same time), we should just pthread_kill the
old thread and join it, that will cause both threads to run as one, and cleanup
will be performed as a unified process, fixing all the refcount issues and
re-register trouble.

You can reproduce the issue by pulling the UTP cable for 3 seconds, shortly after a device
registered. It will take a java phone about 40-50 seconds to start registering again, at this
time the server will have timeout the connection as well, when both coincide -> mayhem.