Fix callforward:Off display update on 89xx Series phones
---
894x phones have a bit of a quirky interface and we are having some trouble adapting
to them. For example when cancelling a callforward the cfwd status is not removed from
the display.

Daljit Parmar provided a pcap which showed a solution to updating the display after
cancelling call forward on the 8945 phones. In retrospect the soluton is pretty simple
and obvious (but contra to how skinny protocol works on other phones). Just go offhook,
cancel callforward and go back onhook.
---

References #465
Fixes #403