Fix for record/monitor feature button on 8945
Apparently the phones firmware contains a bug, displaying the color of the light as the label.
Setting the label in the dynamic text does not have any influence over the outcome on the display.
Switching the 89xx series to old feature_monitor button style and writing extra information after
the label, to show state.

Fixes: #507

Signed-off-by: Diederik de Groot <ddegroot@talon.nl>