Fix use-after-free in sccp_config
If the 'sccp reload line' command is applied multiple times, it will fail due to a use-after-free issue.
This is because sccp_config_getConfig is called with filename pointing to GLOB(config_file_name), so
as a result, newfilename points there. Later, GLOB(config_file_name) is freed and newfilename is
duplicated and placed in GLOB(config_file_name), so GLOB(config_file_name) ends up holding garbage.

Added an inequality check, since in addition to avoiding the use-after-free bug, might as well avoid
doing sccp_free and pbx_strdup for no reason.